requirements.txt
redis
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cryptography==2.8
idna==2.9
jwt==0.6.1
mysql-connector==2.2.9
mysql-connector-python-rf==2.2.2
pkg-resources==0.0.0
pycparser==2.19
python-dotenv==0.12.0
requests==2.23.0
six==1.14.0
urllib3==1.25.8
